requirements.txt
python-snap7
streamlit
pandas
plotly

